backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
faiss-cpu
sentence-transformers
numpy
torch